Big refactor for more idiomatic cap 3 support
