--- lib/capistrano-nginx-unicorn.rb
@@ -1,2 +1,0 @@
-require "capistrano/nginx_unicorn/version"
-require "capistrano/nginx_unicorn/tasks"
