+++ lib/capistrano/nginx_unicorn.rb
@@ -1,0 +1,2 @@
+load File.expand_path("../tasks/nginx.rake", __FILE__)
+load File.expand_path("../tasks/unicorn.rake", __FILE__)
